+++ changelog.md
@@ -5,3 +5,8 @@
 - added git origin, changelog and other git files
 - baseline train ops for contemplating training using class based ops
 - [WIP] need to add save_images_at and save_checkpoint at functions in train ops
+
+### 2020-12-17
+- added author information
+- added implementation for saving and generation of image ops
+- sorted imports